implement changes for better seo
