+++ packages/app/src/embed/components/Sidebar/index.js
@@ -1,15 +1,13 @@
 import * as React from 'react';
 import type { Sandbox } from '@codesandbox/common/lib/types';
 
-import Margin from '@codesandbox/common/lib/components/spacing/Margin';
-import { sandboxUrl } from '@codesandbox/common/lib/utils/url-generator';
 import Section from './Section';
 import { SandboxInfo } from './SandboxInfo';
 import FileTree from './FileTree';
 import Dependencies from './Dependencies';
 import ExternalResources from './ExternalResources';
 
-import { Container, Button } from './elements';
+import { Container } from './elements';
 
 type Props = {
   sandbox: Sandbox,
@@ -23,6 +21,7 @@ function Sidebar({ sandbox, setCurrentModule, currentModule }: Props) {
       <Section title="CodeSandbox" openByDefault>
         <SandboxInfo sandbox={sandbox} />
       </Section>
+
       <Section title="Files" openByDefault>
         <FileTree
           sandbox={sandbox}
@@ -30,7 +29,7 @@ function Sidebar({ sandbox, setCurrentModule, currentModule }: Props) {
           setCurrentModuleId={setCurrentModule}
         />
       </Section>
-      <Section title="Dependencies">
+      <Section title="Dependencies" openByDefault>
         <Dependencies sandbox={sandbox} />
       </Section>
       <Section
@@ -39,16 +38,6 @@ function Sidebar({ sandbox, setCurrentModule, currentModule }: Props) {
       >
         <ExternalResources sandbox={sandbox} />
       </Section>
-
-      <Margin horizontal={0.5} vertical={1}>
-        <Button
-          href={sandboxUrl(sandbox) + '?from-embed'}
-          target="_blank"
-          rel="noreferrer noopener"
-        >
-          Edit Sandbox
-        </Button>
-      </Margin>
     </Container>
   );
 }
